DataTile as a container class
...not as a subclass of an array type.

This avoids several calls to the array constructor, which brings a
noticeable improvement in the inner processing loop for sparse.pydata.org arrays.

Since the tile is unwrapped and the underlying array is passed to the
UDF now, access to the scheme_idx and tile_slice attributes of a tile
is not possible anymore. They are now available under UDFMeta. An
error message points the user to the new alternative.

Since this was a "half-internal" interface, it is trivial to migrate
and is only useful for advanced UDFs, it is not deprecated but removed
hard.